Move compactRelatedBlocks from Page to DictionaryBlock
